Part01
Se adicon asolucion a Part01

+++ src/test/java/io/pivotal/literx/Part01FluxTest.java
@@ -57,9 +57,9 @@ public class Part01FluxTest {
 
 	@Test
 	public void countEach100ms() {
-		Flux<Long> flux = workshop.counter();
+		Flux<Integer> flux = workshop.counter();
 		StepVerifier.create(flux)
-				.expectNext(0L, 1L, 2L, 3L, 4L, 5L, 6L, 7L, 8L, 9L)
+				.expectNext(0, 1, 2, 3, 4, 5, 6, 7, 8, 9)
 				.verifyComplete();
 	}
